decoupled linear from qwen omni specific modules
